+++ 3_types.ts
@@ -47,27 +47,27 @@ export * from "./types/2_input_message_content.ts";
 export * from "./types/2_poll_answer.ts";
 export * from "./types/2_reply_keyboard_markup.ts";
 export * from "./types/3_inline_keyboard_markup.ts";
-export * from "./types/3_inline_query_result_article.ts";
-export * from "./types/3_inline_query_result_audio.ts";
-export * from "./types/3_inline_query_result_cached_audio.ts";
-export * from "./types/3_inline_query_result_cached_document.ts";
-export * from "./types/3_inline_query_result_cached_gif.ts";
-export * from "./types/3_inline_query_result_cached_mpeg4_gif.ts";
-export * from "./types/3_inline_query_result_cached_photo.ts";
-export * from "./types/3_inline_query_result_cached_sticker.ts";
-export * from "./types/3_inline_query_result_cached_video.ts";
-export * from "./types/3_inline_query_result_cached_voice.ts";
-export * from "./types/3_inline_query_result_contact.ts";
-export * from "./types/3_inline_query_result_document.ts";
-export * from "./types/3_inline_query_result_game.ts";
-export * from "./types/3_inline_query_result_gif.ts";
-export * from "./types/3_inline_query_result_location.ts";
-export * from "./types/3_inline_query_result_mpeg4_gif.ts";
-export * from "./types/3_inline_query_result_photo.ts";
-export * from "./types/3_inline_query_result_venue.ts";
-export * from "./types/3_inline_query_result_video.ts";
-export * from "./types/3_inline_query_result_voice.ts";
 export * from "./types/3_message.ts";
 export * from "./types/4_callback_query.ts";
+export * from "./types/4_inline_query_result_article.ts";
+export * from "./types/4_inline_query_result_audio.ts";
+export * from "./types/4_inline_query_result_cached_audio.ts";
+export * from "./types/4_inline_query_result_cached_document.ts";
+export * from "./types/4_inline_query_result_cached_gif.ts";
+export * from "./types/4_inline_query_result_cached_mpeg4_gif.ts";
+export * from "./types/4_inline_query_result_cached_photo.ts";
+export * from "./types/4_inline_query_result_cached_sticker.ts";
+export * from "./types/4_inline_query_result_cached_video.ts";
+export * from "./types/4_inline_query_result_cached_voice.ts";
+export * from "./types/4_inline_query_result_contact.ts";
+export * from "./types/4_inline_query_result_document.ts";
+export * from "./types/4_inline_query_result_game.ts";
+export * from "./types/4_inline_query_result_gif.ts";
+export * from "./types/4_inline_query_result_location.ts";
+export * from "./types/4_inline_query_result_mpeg4_gif.ts";
+export * from "./types/4_inline_query_result_photo.ts";
+export * from "./types/4_inline_query_result_venue.ts";
+export * from "./types/4_inline_query_result_video.ts";
+export * from "./types/4_inline_query_result_voice.ts";
 export * from "./types/4_reply_markup.ts";
 export * from "./types/5_inline_query_result.ts";
