+++ types/4_inline_query_result_audio.ts
@@ -1,17 +1,18 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedVideo {
-  type: "video";
+export interface InlineQueryResultAudio {
+  type: "audio";
   id: string;
-  videoFileId: string;
+  audioUrl: string;
   title: string;
-  description?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities: MessageEntity[];
+  performer?: string;
+  audioDuration?: number;
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
