+++ types/4_inline_query_result_cached_audio.ts
@@ -1,20 +1,15 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultPhoto {
-  type: "photo";
+export interface InlineQueryResultCachedAudio {
+  type: "audio";
   id: string;
-  photoUrl: string;
-  thumbnailUrl: string;
-  photoWidth?: number;
-  photoHeight?: number;
-  title?: string;
-  description?: string;
+  audioFileId: string;
   caption?: string;
   parseMode?: ParseMode;
-  captionEntities?: MessageEntity[];
+  captionEntities: MessageEntity[];
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
