+++ types/4_inline_query_result_cached_document.ts
@@ -1,18 +1,16 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultAudio {
-  type: "audio";
+export interface InlineQueryResultCachedDocument {
+  type: "document";
   id: string;
-  audioUrl: string;
-  title: string;
+  documentFileId: string;
+  description?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities: MessageEntity[];
-  performer?: string;
-  audioDuration?: number;
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
