+++ types/4_inline_query_result_cached_gif.ts
@@ -1,12 +1,13 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedAudio {
-  type: "audio";
+export interface InlineQueryResultCachedGif {
+  type: "gif";
   id: string;
-  audioFileId: string;
+  gifFileId: string;
+  title?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities: MessageEntity[];
