+++ types/4_inline_query_result_cached_mpeg4_gif.ts
@@ -1,7 +1,7 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
 export interface InlineQueryResultCachedMpeg4Gif {
   type: "mpeg4_gif";
