+++ types/4_inline_query_result_cached_photo.ts
@@ -1,12 +1,14 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedDocument {
-  type: "document";
+export interface InlineQueryResultCachedPhoto {
+  type: "photo";
   id: string;
-  documentFileId: string;
+
+  photoFileId: string;
+  title?: string;
   description?: string;
   caption?: string;
   parseMode?: ParseMode;
