+++ types/4_inline_query_result_cached_sticker.ts
@@ -1,5 +1,5 @@
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
 export interface InlineQueryResultCachedSticker {
   type: "sticker";
