+++ types/4_inline_query_result_cached_video.ts
@@ -1,13 +1,14 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedVoice {
-  type: "voice";
+export interface InlineQueryResultCachedVideo {
+  type: "video";
   id: string;
-  voiceFileId: string;
+  videoFileId: string;
   title: string;
+  description?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities: MessageEntity[];
