+++ types/4_inline_query_result_document.ts
@@ -1,18 +1,19 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedPhoto {
-  type: "photo";
+export interface InlineQueryResultDocument {
+  type: "document";
   id: string;
-
-  photoFileId: string;
-  title?: string;
-  description?: string;
+  documentUrl: string;
+  title: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities: MessageEntity[];
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
+  thumbnailUrl?: string;
+  thumbnailWidth?: string;
+  thumbnailHeight?: string;
 }
