+++ types/4_inline_query_result_mpeg4_gif.ts
@@ -1,16 +1,21 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultCachedGif {
-  type: "gif";
+export interface InlineQueryResultMpeg4Gif {
+  type: "mpeg4_gif";
   id: string;
-  gifFileId: string;
+  mpeg4Url: string;
+  mpeg4Width?: number;
+  mpeg4Height?: number;
+  mpeg4Duration?: number;
+  thumbnailUrl?: string;
+  thumbnailMimeType?: string;
   title?: string;
   caption?: string;
   parseMode?: ParseMode;
-  captionEntities: MessageEntity[];
+  captionEntities?: MessageEntity[];
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
