+++ types/4_inline_query_result_photo.ts
@@ -1,22 +1,20 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultVideo {
-  type: "video";
+export interface InlineQueryResultPhoto {
+  type: "photo";
   id: string;
-  videoUrl: string;
-  mimeType?: string;
-  thumbnailUrl?: string;
+  photoUrl: string;
+  thumbnailUrl: string;
+  photoWidth?: number;
+  photoHeight?: number;
   title?: string;
+  description?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities?: MessageEntity[];
-  videoWidth?: number;
-  videoHeight?: number;
-  videoDuration?: number;
-  description?: string;
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
