+++ types/4_inline_query_result_video.ts
@@ -1,21 +1,22 @@
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineKeyboardMarkup } from "./2_inline_keyboard_markup.ts";
 import { InputMessageContent } from "./2_input_message_content.ts";
+import { InlineKeyboardMarkup } from "./3_inline_keyboard_markup.ts";
 
-export interface InlineQueryResultMpeg4Gif {
-  type: "mpeg4_gif";
+export interface InlineQueryResultVideo {
+  type: "video";
   id: string;
-  mpeg4Url: string;
-  mpeg4Width?: number;
-  mpeg4Height?: number;
-  mpeg4Duration?: number;
+  videoUrl: string;
+  mimeType?: string;
   thumbnailUrl?: string;
-  thumbnailMimeType?: string;
   title?: string;
   caption?: string;
   parseMode?: ParseMode;
   captionEntities?: MessageEntity[];
+  videoWidth?: number;
+  videoHeight?: number;
+  videoDuration?: number;
+  description?: string;
   replyMarkup?: InlineKeyboardMarkup;
   inputMessageContent?: InputMessageContent;
 }
