+++ types/5_inline_query_result.ts
@@ -2,28 +2,28 @@ import { types } from "../2_tl.ts";
 import { FileID } from "./0__file_id.ts";
 import { MessageEntity } from "./0_message_entity.ts";
 import { ParseMode } from "./0_parse_mode.ts";
-import { InlineQueryResultArticle } from "./3_inline_query_result_article.ts";
-import { InlineQueryResultAudio } from "./3_inline_query_result_audio.ts";
-import { InlineQueryResultCachedAudio } from "./3_inline_query_result_cached_audio.ts";
-import { InlineQueryResultCachedDocument } from "./3_inline_query_result_cached_document.ts";
-import { InlineQueryResultCachedGif } from "./3_inline_query_result_cached_gif.ts";
-import { InlineQueryResultCachedMpeg4Gif } from "./3_inline_query_result_cached_mpeg4_gif.ts";
-import { InlineQueryResultCachedPhoto } from "./3_inline_query_result_cached_photo.ts";
-import { InlineQueryResultCachedSticker } from "./3_inline_query_result_cached_sticker.ts";
-import { InlineQueryResultCachedVideo } from "./3_inline_query_result_cached_video.ts";
-import { InlineQueryResultCachedVoice } from "./3_inline_query_result_cached_voice.ts";
-import { InlineQueryResultContact } from "./3_inline_query_result_contact.ts";
-import { InlineQueryResultDocument } from "./3_inline_query_result_document.ts";
-import { InlineQueryResultGame } from "./3_inline_query_result_game.ts";
-import { InlineQueryResultGif } from "./3_inline_query_result_gif.ts";
-import { InlineQueryResultLocation } from "./3_inline_query_result_location.ts";
-import { InlineQueryResultMpeg4Gif } from "./3_inline_query_result_mpeg4_gif.ts";
-import { InlineQueryResultPhoto } from "./3_inline_query_result_photo.ts";
-import { InlineQueryResultVenue } from "./3_inline_query_result_venue.ts";
-import { InlineQueryResultVideo } from "./3_inline_query_result_video.ts";
-import { InlineQueryResultVoice } from "./3_inline_query_result_voice.ts";
 import { InputBotInlineMessageMediaAuto } from "../tl/2_types.ts";
 import { UNREACHABLE } from "../1_utilities.ts";
+import { InlineQueryResultArticle } from "./4_inline_query_result_article.ts";
+import { InlineQueryResultAudio } from "./4_inline_query_result_audio.ts";
+import { InlineQueryResultCachedAudio } from "./4_inline_query_result_cached_audio.ts";
+import { InlineQueryResultCachedDocument } from "./4_inline_query_result_cached_document.ts";
+import { InlineQueryResultCachedGif } from "./4_inline_query_result_cached_gif.ts";
+import { InlineQueryResultCachedMpeg4Gif } from "./4_inline_query_result_cached_mpeg4_gif.ts";
+import { InlineQueryResultCachedPhoto } from "./4_inline_query_result_cached_photo.ts";
+import { InlineQueryResultCachedSticker } from "./4_inline_query_result_cached_sticker.ts";
+import { InlineQueryResultCachedVideo } from "./4_inline_query_result_cached_video.ts";
+import { InlineQueryResultCachedVoice } from "./4_inline_query_result_cached_voice.ts";
+import { InlineQueryResultContact } from "./4_inline_query_result_contact.ts";
+import { InlineQueryResultDocument } from "./4_inline_query_result_document.ts";
+import { InlineQueryResultGame } from "./4_inline_query_result_game.ts";
+import { InlineQueryResultGif } from "./4_inline_query_result_gif.ts";
+import { InlineQueryResultLocation } from "./4_inline_query_result_location.ts";
+import { InlineQueryResultMpeg4Gif } from "./4_inline_query_result_mpeg4_gif.ts";
+import { InlineQueryResultPhoto } from "./4_inline_query_result_photo.ts";
+import { InlineQueryResultVenue } from "./4_inline_query_result_venue.ts";
+import { InlineQueryResultVideo } from "./4_inline_query_result_video.ts";
+import { InlineQueryResultVoice } from "./4_inline_query_result_voice.ts";
 import { replyMarkupToTlObject } from "./4_reply_markup.ts";
 import { UsernameResolver } from "./1__getters.ts";
